# services/user-service/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse'
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.PasswordResetRequestRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ProfileResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      profile_id:
        type: string
      profile_picture_url:
        type: string
      register_no:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.SessionResponse:
    properties:
      created_at:
        type: string
      device_info:
        type: string
      expires_at:
        type: string
      ip_address:
        type: string
      session_id:
        type: string
    type: object
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      last_login:
        type: string
      profile:
        $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ProfileResponse'
      register_no:
        type: integer
      role_id:
        type: string
      role_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@nimbusu.edu
    name: API Support
    url: http://www.nimbusu.edu/support
  description: User Management Service for NimbusU University Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NimbusU User Service API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access and refresh tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: User Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate current session and refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - auth
  /auth/password/change:
    post:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password Change Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - auth
  /auth/password/reset:
    post:
      consumes:
      - application/json
      description: Reset password using a valid token
      parameters:
      - description: Reset Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Reset Password
      tags:
      - auth
  /auth/password/reset-request:
    post:
      consumes:
      - application/json
      description: Request a password reset email (generates token)
      parameters:
      - description: Email Address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.PasswordResetRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Request Password Reset
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.RefreshTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/sessions:
    delete:
      consumes:
      - application/json
      description: Revoke all active sessions for the current user (logout everywhere)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke All Sessions
      tags:
      - auth
    get:
      consumes:
      - application/json
      description: Retrieve a list of all active sessions for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.SessionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Active Sessions
      tags:
      - auth
  /auth/sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific session by ID
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke Session
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"